pull out config
