+++ pymedphys/_gui/streamlit/mudensity/main.py
@@ -67,96 +67,6 @@ GRID = pymedphys.mudensity.grid(
 COORDS = (GRID["jaw"], GRID["mlc"])
 
 
-@st.cache
-def get_dicom_export_locations():
-    site_directories = st_config.get_site_directories()
-    dicom_export_locations = {
-        site: directories["monaco"].parent.parent.joinpath("DCMXprtFile")
-        for site, directories in site_directories.items()
-    }
-
-    return dicom_export_locations
-
-
-@st.cache
-def get_icom_live_stream_directories():
-    config = st_config.get_config()
-    icom_live_stream_directories = {}
-    for site in config["site"]:
-        icom_live_base_directory = pathlib.Path(site["export-directories"]["icom_live"])
-        for linac in site["linac"]:
-            icom_live_stream_directories[linac["name"]] = str(
-                icom_live_base_directory.joinpath(linac["ip"])
-            )
-
-    return icom_live_stream_directories
-
-
-@st.cache
-def get_machine_centre_map():
-    config = st_config.get_config()
-    machine_centre_map = {}
-    for site in config["site"]:
-        for linac in site["linac"]:
-            machine_centre_map[linac["name"]] = site["name"]
-
-    return machine_centre_map
-
-
-@st.cache
-def get_logfile_root_dir():
-    config = st_config.get_config()
-    logfile_root_dir = pathlib.Path(config["trf_logfiles"]["root_directory"])
-
-    return logfile_root_dir
-
-
-@st.cache
-def get_indexed_backups_directory():
-    logfile_root_dir = get_logfile_root_dir()
-    indexed_backups_directory = logfile_root_dir.joinpath("diagnostics/already_indexed")
-
-    return indexed_backups_directory
-
-
-@st.cache
-def get_indexed_trf_directory():
-    logfile_root_dir = get_logfile_root_dir()
-    indexed_trf_directory = logfile_root_dir.joinpath("indexed")
-
-    return indexed_trf_directory
-
-
-@st.cache
-def get_mosaiq_details():
-    config = st_config.get_config()
-    mosaiq_details = {
-        site["name"]: {
-            "timezone": site["mosaiq"]["timezone"],
-            "server": f'{site["mosaiq"]["hostname"]}:{site["mosaiq"]["port"]}',
-        }
-        for site in config["site"]
-    }
-
-    return mosaiq_details
-
-
-@st.cache
-def get_default_icom_directories():
-    config = st_config.get_config()
-    default_icom_directory = config["icom"]["patient_directories"]
-
-    return default_icom_directory
-
-
-@st.cache
-def get_default_gamma_options():
-    config = st_config.get_config()
-    default_gamma_options = config["gamma"]
-
-    return default_gamma_options
-
-
 class InputRequired(ValueError):
     pass
 
@@ -243,40 +153,6 @@ def show_status_indicators():
     """
     ## Selection of data to compare
     """
-
-
-def get_gamma_options(advanced_mode_local):
-    default_gamma_options = get_default_gamma_options()
-
-    if advanced_mode_local:
-        st.sidebar.markdown(
-            """
-            # Gamma parameters
-            """
-        )
-        result = {
-            **default_gamma_options,
-            **{
-                "dose_percent_threshold": st.sidebar.number_input(
-                    "MU Percent Threshold",
-                    value=default_gamma_options["dose_percent_threshold"],
-                ),
-                "distance_mm_threshold": st.sidebar.number_input(
-                    "Distance (mm) Threshold",
-                    value=default_gamma_options["distance_mm_threshold"],
-                ),
-                "local_gamma": st.sidebar.checkbox(
-                    "Local Gamma", default_gamma_options["local_gamma"]
-                ),
-                "max_gamma": st.sidebar.number_input(
-                    "Max Gamma", value=default_gamma_options["max_gamma"]
-                ),
-            },
-        }
-    else:
-        result = default_gamma_options
-
-    return result
 
 
 @st.cache(allow_output_mutation=True)
@@ -708,41 +584,6 @@ def icom_input_method(patient_id="", key_namespace="", advanced_mode_local=False
 @st.cache
 def read_trf(filepath):
     return pymedphys.read_trf(filepath)
-
-
-@st.cache
-def get_logfile_mosaiq_info(headers):
-    machine_centre_map = get_machine_centre_map()
-    mosaiq_details = get_mosaiq_details()
-
-    centres = {machine_centre_map[machine_id] for machine_id in headers["machine"]}
-    mosaiq_servers = [mosaiq_details[centre]["server"] for centre in centres]
-
-    details = []
-
-    cursors = {server: st_mosaiq.get_mosaiq_cursor(server) for server in mosaiq_servers}
-
-    for _, header in headers.iterrows():
-        machine_id = header["machine"]
-        centre = machine_centre_map[machine_id]
-        mosaiq_timezone = mosaiq_details[centre]["timezone"]
-        server = mosaiq_details[centre]["server"]
-        cursor = cursors[server]
-
-        field_label = header["field_label"]
-        field_name = header["field_name"]
-        utc_date = header["date"]
-
-        current_details = pmp_index.get_logfile_mosaiq_info(
-            cursor, machine_id, utc_date, mosaiq_timezone, field_label, field_name
-        )
-        current_details = pd.Series(data=current_details)
-
-        details.append(current_details)
-
-    details = pd.concat(details, axis=1).T
-
-    return details
 
 
 def trf_input_method(patient_id="", key_namespace="", **_):
